use self.dtype

+++ tests/chainer_tests/functions_tests/loss_tests/test_ctc.py
@@ -16,7 +16,7 @@ from chainer.testing import condition
 class CTCTestBase(object):
 
     def setUp(self):
-        self.x = numpy.random.uniform(-1, 1, (4, 2, 3)).astype(numpy.float32)
+        self.x = numpy.random.uniform(-1, 1, (4, 2, 3)).astype(self.dtype)
         self.t = numpy.array([[2, 1], [1, 2]]).astype(numpy.int32)
         self.l = numpy.array([[0, 2, 0, 1, 0],
                               [0, 1, 0, 2, 0]]).astype(numpy.int32)
@@ -233,7 +233,7 @@ class TestCTCBlankSymbol(unittest.TestCase, CTCTestBase):
 
     def setUp(self):
         CTCTestBase.setUp(self)
-        self.x = numpy.random.uniform(-1, 1, (4, 2, 4)).astype(numpy.float32)
+        self.x = numpy.random.uniform(-1, 1, (4, 2, 4)).astype(self.dtype)
         self.l = numpy.array([[0, 3, 0, 1, 0],
                               [0, 1, 0, 3, 0]]).astype(numpy.int32)
         self.t = numpy.array([[3, 1], [1, 3]]).astype(numpy.int32)
